+++ NEWS.md
@@ -4,3 +4,5 @@
 * Added `R/theme.R`,
     + `theme_mc()`, a theme with `c("#333333", "#7F7F7F", "#FAFAFA")` palette from background to foreground.
     + `theme_mc_md()`, `theme_mc()` with markdown ability using `ggtext` (if installed).
+* Added  `R/check_commit.R`,
+    +  `check_commiy()`, a function to recursively checks uncommitted files in subdirectories.